Avoid rerender the entire components tree
